pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.99.5",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
